requirements.txt
opencv-python==4.8.0.74
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
pillow==10.0.0
tqdm==4.65.0
